requirements.txt
Flask==3.0.0
Flask-Limiter==4.0.0
Flask-WTF==1.2.2
requests==2.32.5
SQLAlchemy==2.0.44
python-dotenv==1.0.0
schedule==1.2.0
iso3166==2.1.1
